+++ packages/hydrogen/src/client.ts
@@ -6,3 +6,4 @@ export * from './foundation/ServerStateProvider';
 export {Head} from './foundation/Head';
 export * from './utilities';
 export {useParams} from './foundation/Router/useParams';
+export {useNavigate} from './foundation/Router/useNavigate';
