+++ packages/hydrogen/src/components/index.ts
@@ -36,4 +36,3 @@ export {ShopPayButton} from './ShopPayButton';
 export {useAvailableCountries} from '../hooks/useAvailableCountries';
 export {useCountry} from '../hooks/useCountry';
 export {Seo} from './Seo';
-export {useNavigate} from '../hooks/useNavigate';
