--- packages/hydrogen/src/hooks/useNavigate/index.ts
@@ -1,1 +1,0 @@
-export {useNavigate} from './useNavigate';
